Achieve buildability on Windows.
The intent is not for this module to *function* on Windows, as that is impossible, but rather for it to not break builds where it may be included as a dependency.

For now, we return `nil` on non-Unix platforms, but in the future we should return an error so that the intention is clearer.